+++ packages/ng-syz/projects/app/src/app/features/components/docs-navbar/docs-navbar.component.html
@@ -1,5 +1,5 @@
 <app-component-page>
     <app-code-example>
-        <ng-syz-app-navbar></ng-syz-app-navbar>
+        <ng-syz-navbar></ng-syz-navbar>
     </app-code-example>
 </app-component-page>
