+++ packages/ng-syz/projects/ng-syz/src/lib/navbar/ng-syz-navbar.component.scss
@@ -1,0 +1,47 @@
+.navbar-spacer {
+    flex: 1 1 auto;
+  }
+  
+  .navbar-item {
+    display: flex;
+    align-items: center;
+  }
+
+  .center{
+    display: flex;
+    align-items: center;
+    margin: auto; 
+  }
+
+  .right{
+    display: flex;
+    align-items: right;
+    margin: auto; 
+  }
+
+  .navbar-brand {
+    margin-left: 10%;
+   }
+
+.mat-toolbar-row,
+.mat-toolbar-single-row {
+  height: 120px;
+  border-bottom: 5px solid var(--primary-color);
+  box-shadow: 0 8px 16px 0 rgba(0,0,0,.26);
+}
+
+button{
+        font-size:large;
+}
+
+button:hover{
+    margin-bottom: solid; 
+    border-bottom:2px solid;
+    
+}
+
+
+
+
+
+
