+++ packages/ng-syz/projects/ng-syz/src/lib/navbar/ng-syz-navbar.component.ts
@@ -1,0 +1,70 @@
+import { Component, OnInit } from '@angular/core';
+ 
+export interface MenuItem {
+  label: string;
+}
+
+export interface MenuButton {
+  icon: string;
+  tooltip: string;
+}
+
+@Component({
+  selector: 'ng-syz-navbar',
+  templateUrl: './ng-syz-navbar.component.html',
+  styleUrls: ['./ng-syz-navbar.component.scss']
+})
+export class NgSyzNavbarComponent implements OnInit {
+
+  backGroundColor = "#FFF";
+
+ menuItems: MenuItem[] = [
+    {
+      label: 'Iníco'
+     
+    },
+    {
+      label: 'Equipe'
+   
+    },
+    {
+      label: 'Performance'
+      
+    },
+    {
+      label: 'Gestão'
+    },
+    {
+      label: 'Calendário'
+    
+    },
+    {
+      label: 'Documentos'
+   
+    },
+    {
+      label: 'Duvidas'
+   
+    }
+  ];
+
+  menuButtons: MenuButton[] = [
+    {
+      icon: 'logout',
+      tooltip: 'sair',
+    },
+    {
+      icon: 'notifications',
+      tooltip:'alerta'
+    },
+
+  ];
+
+  
+
+  constructor() { }
+ 
+  ngOnInit(): void {
+  }
+ 
+}
